Moved `AzureSSOConfiguration` to its own class
