+++ src/Our.Umbraco.AzureSSO/AzureSSOConfiguration.cs
@@ -1,0 +1,22 @@
+using System.Collections.Generic;
+
+namespace Our.Umbraco.AzureSSO
+{
+	public class AzureSSOConfiguration
+	{
+		public AzureSSOConfiguration()
+		{
+			GroupBindings = new Dictionary<string, string>();
+		}
+
+		public string? DisplayName { get; set; }
+
+		public string? ButtonStyle { get; set; }
+
+		public string? Icon { get; set; }
+
+		public Dictionary<string, string> GroupBindings { get; set; }
+
+		public bool? DenyLocalLogin { get; set; }
+	}
+}
