+++ src/Our.Umbraco.AzureSSO/MicrosoftAccountAuthenticationExtensions.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
 using Microsoft.AspNetCore.Authentication.OpenIdConnect;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
@@ -43,20 +42,6 @@ namespace Our.Umbraco.AzureSSO
 			});
 			return builder;
 		}
-		
+
 	}
-
-	public class AzureSSOConfiguration
-	{
-		public AzureSSOConfiguration()
-		{
-			GroupBindings = new Dictionary<string, string>();
-		}
-
-		public string? DisplayName { get; set; }
-		public string? ButtonStyle { get; set; }
-		public string? Icon { get; set; }
-		public Dictionary<string,string> GroupBindings { get; set; }
-		public bool? DenyLocalLogin { get; set; }
- 	}
 }
